Extract CoastalVoyageModel definition into its own file
